add rest api module and tests
