+++ examples/account-settings/main.tf
@@ -1,3 +1,7 @@
+provider "aws" {
+  region = var.region
+}
+
 module "example" {
   source  = "../../modules/account-settings"
   context = module.this.context
