+++ examples/rest-api/fixtures.us-east-2.tfvars
@@ -1,0 +1,4 @@
+region      = "us-east-2"
+namespace   = "eg"
+environment = "ue2"
+stage       = "test"
