+++ examples/rest-api/main.tf
@@ -1,0 +1,28 @@
+provider "aws" {
+  region = var.region
+}
+
+module "example" {
+  source = "../../"
+
+  openapi_config = {
+    openapi = "3.0.1"
+    info = {
+      title   = "example"
+      version = "1.0"
+    }
+    paths = {
+      "/path1" = {
+        get = {
+          x-amazon-apigateway-integration = {
+            httpMethod           = "GET"
+            payloadFormatVersion = "1.0"
+            type                 = "HTTP_PROXY"
+            uri                  = "https://ip-ranges.amazonaws.com/ip-ranges.json"
+          }
+        }
+      }
+    }
+  }
+  context = module.this.context
+}
