+++ examples/rest-api/outputs.tf
@@ -1,0 +1,9 @@
+output "created_date" {
+  description = "The date the REST API was created"
+  value       = module.example.created_date
+}
+
+output "invoke_url" {
+  description = "The URL to invoke the REST API"
+  value = module.example.invoke_url
+}
