+++ examples/rest-api/variables.tf
@@ -1,0 +1,4 @@
+variable "region" {
+  type        = string
+  description = "AWS Region"
+}
