+++ examples/rest-api/versions.tf
@@ -1,0 +1,10 @@
+terraform {
+  required_version = ">= 0.13"
+
+  required_providers {
+    local = {
+      source  = "hashicorp/aws"
+      version = ">= 3.0"
+    }
+  }
+}
