+++ test/src/examples_account_settings_test.go
@@ -43,3 +43,31 @@ func TestExamplesAccountSettings(t *testing.T) {
 	// Ensure we get a random number appended
 	assert.NotNil(t, role_arn)
 }
+
+// Test the Terraform module in examples/account-settings doesn't attenmpt to create resources with enabled=false.
+func TestExamplesAccountSettingsDisabled(t *testing.T) {
+	t.Parallel()
+
+	rand.Seed(time.Now().UnixNano())
+	randID := strconv.Itoa(rand.Intn(100000))
+	attributes := []string{randID}
+
+	terraformOptions := terraform.WithDefaultRetryableErrors(t, &terraform.Options{
+		// The path to where our Terraform code is located
+		TerraformDir: "../../examples/account-settings",
+		Upgrade:      true,
+		// Variables to pass to our Terraform code using -var-file options
+		VarFiles: []string{"fixtures.us-east-2.tfvars"},
+		// We always include a random attribute so that parallel tests
+		// and AWS resources do not interfere with each other
+		Vars: map[string]interface{}{
+			"enabled":    false,
+			"attributes": attributes,
+		},
+	})
+
+	terraform.Init(t, terraformOptions)
+	plan := terraform.Plan(t, terraformOptions)
+
+	assert.Contains(t, plan, "No changes. Infrastructure is up-to-date")
+}
