+++ test/src/go.mod
@@ -3,10 +3,6 @@ module github.com/cloudposse/terraform-example-module
 go 1.13
 
 require (
-	github.com/gruntwork-io/terratest v0.28.15
-	github.com/stretchr/testify v1.6.1
-	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a // indirect
-	golang.org/x/net v0.0.0-20200822124328-c89045814202 // indirect
-	golang.org/x/sys v0.0.0-20200828194041-157a740278f4 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776 // indirect
+	github.com/gruntwork-io/terratest v0.38.6
+	github.com/stretchr/testify v1.7.0
 )
