+++ variables.tf
@@ -1,0 +1,67 @@
+variable "openapi_config" {
+  type    = any
+  default = {}
+}
+
+variable "endpoint_type" {
+  type    = string
+  default = "REGIONAL"
+
+  validation {
+    condition     = contains(["EDGE", "REGIONAL", "PRIVATE"], var.endpoint_type)
+    error_message = "Valid values for var: endpoint_type are (EDGE, REGIONAL, PRIVATE)."
+  }
+}
+
+variable "logging_level" {
+  type    = string
+  default = "INFO"
+
+  validation {
+    condition     = contains(["OFF", "INFO", "ERROR"], var.logging_level)
+    error_message = "Valid values for var: logging_level are (OFF, INFO, ERROR)."
+  }
+}
+
+variable "metrics_enabled" {
+  description = "A flag to indicate whether to enable metrics collection."
+  type        = bool
+  default     = false
+}
+
+variable "xray_tracing_enabled" {
+  description = "A flag to indicate whether to enable X-Ray tracing."
+  type        = bool
+  default     = false
+}
+
+variable "access_log_format" {
+  description = "The format of the access log file."
+  type        = string
+  default     = <<EOF
+  {
+	"requestTime": "$context.requestTime",
+	"requestId": "$context.requestId",
+	"httpMethod": "$context.httpMethod",
+	"path": "$context.path",
+	"resourcePath": "$context.resourcePath",
+	"status": $context.status,
+	"responseLatency": $context.responseLatency,
+  "xrayTraceId": "$context.xrayTraceId",
+  "integrationRequestId": "$context.integration.requestId",
+	"functionResponseStatus": "$context.integration.status",
+  "integrationLatency": "$context.integration.latency",
+	"integrationServiceStatus": "$context.integration.integrationStatus",
+  "authorizeResultStatus": "$context.authorize.status",
+	"authorizerServiceStatus": "$context.authorizer.status",
+	"authorizerLatency": "$context.authorizer.latency",
+	"authorizerRequestId": "$context.authorizer.requestId",
+  "ip": "$context.identity.sourceIp",
+	"userAgent": "$context.identity.userAgent",
+	"principalId": "$context.authorizer.principalId",
+	"cognitoUser": "$context.identity.cognitoIdentityId",
+  "user": "$context.identity.user"
+}
+  EOF
+}
+
